requirements.txt
numpy
google-genai
python-dotenv
pypdf
python-docx
tqdm
